Separated WASM module into its own file.
