+++ ripemd160.mjs
@@ -1,348 +1,5 @@
 // @ts-check
-
-const WASM = [
-  0, 97, 115, 109, 1, 0, 0, 0, 1, 15, 3, 96, 1, 127, 0, 96, 1, 127, 1, 127, 96,
-  2, 127, 127, 0, 3, 6, 5, 1, 0, 2, 1, 0, 5, 3, 1, 0, 2, 6, 8, 1, 127, 1, 65,
-  176, 137, 4, 11, 7, 29, 3, 6, 109, 101, 109, 111, 114, 121, 2, 0, 7, 103, 101,
-  116, 95, 112, 116, 114, 0, 3, 6, 100, 105, 103, 101, 115, 116, 0, 4, 10, 165,
-  48, 5, 37, 1, 1, 127, 3, 64, 32, 1, 65, 192, 0, 71, 4, 64, 32, 0, 32, 1, 106,
-  65, 0, 58, 0, 0, 32, 1, 65, 1, 106, 33, 1, 12, 1, 11, 11, 32, 0, 11, 211, 3,
-  1, 8, 127, 65, 192, 8, 33, 4, 35, 0, 65, 224, 0, 107, 34, 3, 36, 0, 32, 3, 65,
-  208, 0, 106, 65, 144, 8, 40, 2, 0, 54, 2, 0, 32, 3, 65, 136, 8, 41, 3, 0, 55,
-  3, 72, 32, 3, 65, 128, 8, 41, 3, 0, 55, 3, 64, 32, 0, 65, 6, 118, 33, 8, 3,
-  64, 32, 5, 32, 8, 71, 4, 64, 65, 0, 33, 1, 3, 64, 32, 1, 32, 4, 106, 33, 2,
-  32, 1, 65, 192, 0, 71, 4, 64, 32, 1, 32, 3, 106, 34, 6, 32, 2, 45, 0, 0, 34,
-  7, 54, 2, 0, 32, 6, 32, 7, 32, 2, 65, 1, 106, 45, 0, 0, 65, 8, 116, 114, 34,
-  7, 54, 2, 0, 32, 6, 32, 2, 65, 2, 106, 45, 0, 0, 65, 16, 116, 32, 7, 114, 34,
-  7, 54, 2, 0, 32, 6, 32, 2, 65, 3, 106, 45, 0, 0, 65, 24, 116, 32, 7, 114, 54,
-  2, 0, 32, 1, 65, 4, 106, 33, 1, 12, 1, 11, 11, 32, 5, 65, 1, 106, 33, 5, 32,
-  3, 65, 64, 107, 32, 3, 16, 2, 32, 2, 33, 4, 12, 1, 11, 11, 65, 0, 33, 5, 32,
-  3, 16, 0, 33, 2, 32, 0, 65, 63, 113, 33, 3, 65, 0, 33, 1, 3, 64, 32, 1, 32, 3,
-  71, 4, 64, 32, 2, 32, 1, 65, 124, 113, 106, 34, 6, 32, 6, 40, 2, 0, 32, 1, 32,
-  4, 106, 45, 0, 0, 32, 5, 65, 24, 113, 116, 115, 54, 2, 0, 32, 5, 65, 8, 106,
-  33, 5, 32, 1, 65, 1, 106, 33, 1, 12, 1, 11, 11, 32, 2, 32, 0, 65, 60, 113,
-  106, 34, 4, 32, 4, 40, 2, 0, 65, 1, 32, 0, 65, 3, 116, 34, 4, 65, 24, 113, 65,
-  7, 114, 116, 115, 54, 2, 0, 32, 3, 65, 56, 79, 4, 64, 32, 2, 65, 64, 107, 32,
-  2, 16, 2, 32, 2, 16, 0, 26, 11, 32, 2, 32, 4, 54, 2, 56, 32, 2, 32, 0, 65, 29,
-  118, 54, 2, 60, 32, 2, 65, 64, 107, 32, 2, 16, 2, 65, 0, 33, 1, 3, 64, 32, 1,
-  65, 20, 71, 4, 64, 32, 1, 65, 160, 8, 106, 34, 0, 32, 2, 65, 64, 107, 32, 1,
-  106, 34, 4, 40, 2, 0, 34, 3, 58, 0, 0, 32, 0, 65, 2, 106, 32, 4, 47, 1, 2, 58,
-  0, 0, 32, 0, 65, 3, 106, 32, 4, 45, 0, 3, 58, 0, 0, 32, 0, 65, 1, 106, 32, 3,
-  65, 8, 118, 58, 0, 0, 32, 1, 65, 4, 106, 33, 1, 12, 1, 11, 11, 32, 2, 65, 224,
-  0, 106, 36, 0, 11, 251, 43, 1, 32, 127, 32, 0, 32, 1, 40, 2, 44, 34, 25, 32,
-  1, 40, 2, 40, 34, 15, 32, 1, 40, 2, 20, 34, 18, 32, 18, 32, 1, 40, 2, 52, 34,
-  26, 32, 15, 32, 18, 32, 1, 40, 2, 28, 34, 20, 32, 1, 40, 2, 36, 34, 27, 32, 1,
-  40, 2, 32, 34, 10, 32, 27, 32, 1, 40, 2, 24, 34, 22, 32, 20, 32, 25, 32, 22,
-  32, 1, 40, 2, 4, 34, 19, 32, 0, 40, 2, 16, 34, 30, 106, 32, 0, 40, 2, 8, 34,
-  31, 65, 10, 119, 34, 5, 32, 0, 40, 2, 4, 34, 29, 115, 32, 1, 40, 2, 0, 34, 23,
-  32, 0, 40, 2, 0, 34, 32, 32, 0, 40, 2, 12, 34, 4, 32, 29, 32, 31, 115, 115,
-  106, 106, 65, 11, 119, 32, 30, 106, 34, 17, 115, 106, 65, 14, 119, 32, 4, 106,
-  34, 16, 65, 10, 119, 34, 3, 106, 32, 1, 40, 2, 16, 34, 21, 32, 29, 65, 10,
-  119, 34, 7, 106, 32, 1, 40, 2, 8, 34, 24, 32, 4, 106, 32, 7, 32, 17, 115, 32,
-  16, 115, 106, 65, 15, 119, 32, 5, 106, 34, 2, 32, 3, 115, 32, 1, 40, 2, 12,
-  34, 28, 32, 5, 106, 32, 16, 32, 17, 65, 10, 119, 34, 17, 115, 32, 2, 115, 106,
-  65, 12, 119, 32, 7, 106, 34, 16, 115, 106, 65, 5, 119, 32, 17, 106, 34, 6, 32,
-  16, 65, 10, 119, 34, 8, 115, 32, 17, 32, 18, 106, 32, 16, 32, 2, 65, 10, 119,
-  34, 17, 115, 32, 6, 115, 106, 65, 8, 119, 32, 3, 106, 34, 16, 115, 106, 65, 7,
-  119, 32, 17, 106, 34, 3, 65, 10, 119, 34, 2, 106, 32, 27, 32, 6, 65, 10, 119,
-  34, 6, 106, 32, 17, 32, 20, 106, 32, 6, 32, 16, 115, 32, 3, 115, 106, 65, 9,
-  119, 32, 8, 106, 34, 17, 32, 2, 115, 32, 8, 32, 10, 106, 32, 3, 32, 16, 65,
-  10, 119, 34, 16, 115, 32, 17, 115, 106, 65, 11, 119, 32, 6, 106, 34, 3, 115,
-  106, 65, 13, 119, 32, 16, 106, 34, 6, 32, 3, 65, 10, 119, 34, 8, 115, 32, 2,
-  32, 15, 32, 16, 106, 32, 3, 32, 17, 65, 10, 119, 34, 2, 115, 32, 6, 115, 106,
-  65, 14, 119, 106, 34, 3, 115, 106, 65, 15, 119, 32, 2, 106, 34, 9, 65, 10,
-  119, 34, 11, 106, 32, 3, 65, 10, 119, 34, 12, 32, 1, 40, 2, 60, 34, 17, 106,
-  32, 8, 32, 26, 106, 32, 2, 32, 1, 40, 2, 48, 34, 16, 106, 32, 3, 32, 6, 65,
-  10, 119, 34, 2, 115, 32, 9, 115, 106, 65, 6, 119, 32, 8, 106, 34, 3, 32, 9,
-  32, 12, 115, 115, 106, 65, 7, 119, 32, 2, 106, 34, 6, 32, 3, 65, 10, 119, 34,
-  8, 115, 32, 2, 32, 1, 40, 2, 56, 34, 1, 106, 32, 3, 32, 11, 115, 32, 6, 115,
-  106, 65, 9, 119, 32, 12, 106, 34, 9, 115, 106, 65, 8, 119, 32, 11, 106, 34, 3,
-  32, 9, 113, 32, 6, 65, 10, 119, 34, 6, 32, 3, 65, 127, 115, 113, 114, 106, 65,
-  153, 243, 137, 212, 5, 106, 65, 7, 119, 32, 8, 106, 34, 2, 65, 10, 119, 34,
-  11, 106, 32, 15, 32, 3, 65, 10, 119, 34, 12, 106, 32, 19, 32, 9, 65, 10, 119,
-  34, 9, 106, 32, 6, 32, 26, 106, 32, 8, 32, 21, 106, 32, 2, 32, 3, 113, 32, 9,
-  32, 2, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65, 6,
-  119, 32, 6, 106, 34, 3, 32, 2, 113, 32, 12, 32, 3, 65, 127, 115, 113, 114,
-  106, 65, 153, 243, 137, 212, 5, 106, 65, 8, 119, 32, 9, 106, 34, 2, 32, 3,
-  113, 32, 11, 32, 2, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5,
-  106, 65, 13, 119, 32, 12, 106, 34, 6, 32, 2, 113, 32, 3, 65, 10, 119, 34, 8,
-  32, 6, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65, 11,
-  119, 32, 11, 106, 34, 3, 32, 6, 113, 32, 2, 65, 10, 119, 34, 9, 32, 3, 65,
-  127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65, 9, 119, 32, 8,
-  106, 34, 2, 65, 10, 119, 34, 11, 106, 32, 23, 32, 3, 65, 10, 119, 34, 12, 106,
-  32, 16, 32, 6, 65, 10, 119, 34, 6, 106, 32, 9, 32, 28, 106, 32, 8, 32, 17,
-  106, 32, 2, 32, 3, 113, 32, 6, 32, 2, 65, 127, 115, 113, 114, 106, 65, 153,
-  243, 137, 212, 5, 106, 65, 7, 119, 32, 9, 106, 34, 3, 32, 2, 113, 32, 12, 32,
-  3, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65, 15, 119,
-  32, 6, 106, 34, 2, 32, 3, 113, 32, 11, 32, 2, 65, 127, 115, 113, 114, 106, 65,
-  153, 243, 137, 212, 5, 106, 65, 7, 119, 32, 12, 106, 34, 6, 32, 2, 113, 32, 3,
-  65, 10, 119, 34, 8, 32, 6, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137,
-  212, 5, 106, 65, 12, 119, 32, 11, 106, 34, 3, 32, 6, 113, 32, 2, 65, 10, 119,
-  34, 9, 32, 3, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65,
-  15, 119, 32, 8, 106, 34, 2, 65, 10, 119, 34, 11, 106, 32, 25, 32, 3, 65, 10,
-  119, 34, 12, 106, 32, 1, 32, 6, 65, 10, 119, 34, 6, 106, 32, 9, 32, 24, 106,
-  32, 8, 32, 18, 106, 32, 2, 32, 3, 113, 32, 6, 32, 2, 65, 127, 115, 113, 114,
-  106, 65, 153, 243, 137, 212, 5, 106, 65, 9, 119, 32, 9, 106, 34, 3, 32, 2,
-  113, 32, 12, 32, 3, 65, 127, 115, 113, 114, 106, 65, 153, 243, 137, 212, 5,
-  106, 65, 11, 119, 32, 6, 106, 34, 2, 32, 3, 113, 32, 11, 32, 2, 65, 127, 115,
-  113, 114, 106, 65, 153, 243, 137, 212, 5, 106, 65, 7, 119, 32, 12, 106, 34, 6,
-  32, 2, 113, 32, 3, 65, 10, 119, 34, 3, 32, 6, 65, 127, 115, 113, 114, 106, 65,
-  153, 243, 137, 212, 5, 106, 65, 13, 119, 32, 11, 106, 34, 8, 32, 6, 113, 32,
-  2, 65, 10, 119, 34, 2, 32, 8, 65, 127, 115, 34, 12, 113, 114, 106, 65, 153,
-  243, 137, 212, 5, 106, 65, 12, 119, 32, 3, 106, 34, 9, 65, 10, 119, 34, 11,
-  106, 32, 21, 32, 8, 65, 10, 119, 34, 8, 106, 32, 1, 32, 6, 65, 10, 119, 34, 6,
-  106, 32, 2, 32, 15, 106, 32, 3, 32, 28, 106, 32, 9, 32, 12, 114, 32, 6, 115,
-  106, 65, 161, 215, 231, 246, 6, 106, 65, 11, 119, 32, 2, 106, 34, 3, 32, 9,
-  65, 127, 115, 114, 32, 8, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 13,
-  119, 32, 6, 106, 34, 2, 32, 3, 65, 127, 115, 114, 32, 11, 115, 106, 65, 161,
-  215, 231, 246, 6, 106, 65, 6, 119, 32, 8, 106, 34, 6, 32, 2, 65, 127, 115,
-  114, 32, 3, 65, 10, 119, 34, 3, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65,
-  7, 119, 32, 11, 106, 34, 8, 32, 6, 65, 127, 115, 114, 32, 2, 65, 10, 119, 34,
-  2, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 14, 119, 32, 3, 106, 34, 9,
-  65, 10, 119, 34, 11, 106, 32, 24, 32, 8, 65, 10, 119, 34, 12, 106, 32, 19, 32,
-  6, 65, 10, 119, 34, 6, 106, 32, 2, 32, 10, 106, 32, 3, 32, 17, 106, 32, 9, 32,
-  8, 65, 127, 115, 114, 32, 6, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 9,
-  119, 32, 2, 106, 34, 3, 32, 9, 65, 127, 115, 114, 32, 12, 115, 106, 65, 161,
-  215, 231, 246, 6, 106, 65, 13, 119, 32, 6, 106, 34, 2, 32, 3, 65, 127, 115,
-  114, 32, 11, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 15, 119, 32, 12,
-  106, 34, 6, 32, 2, 65, 127, 115, 114, 32, 3, 65, 10, 119, 34, 3, 115, 106, 65,
-  161, 215, 231, 246, 6, 106, 65, 14, 119, 32, 11, 106, 34, 8, 32, 6, 65, 127,
-  115, 114, 32, 2, 65, 10, 119, 34, 2, 115, 106, 65, 161, 215, 231, 246, 6, 106,
-  65, 8, 119, 32, 3, 106, 34, 9, 65, 10, 119, 34, 11, 106, 32, 25, 32, 8, 65,
-  10, 119, 34, 12, 106, 32, 26, 32, 6, 65, 10, 119, 34, 6, 106, 32, 2, 32, 22,
-  106, 32, 3, 32, 23, 106, 32, 9, 32, 8, 65, 127, 115, 114, 32, 6, 115, 106, 65,
-  161, 215, 231, 246, 6, 106, 65, 13, 119, 32, 2, 106, 34, 3, 32, 9, 65, 127,
-  115, 114, 32, 12, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 6, 119, 32, 6,
-  106, 34, 2, 32, 3, 65, 127, 115, 114, 32, 11, 115, 106, 65, 161, 215, 231,
-  246, 6, 106, 65, 5, 119, 32, 12, 106, 34, 6, 32, 2, 65, 127, 115, 114, 32, 3,
-  65, 10, 119, 34, 8, 115, 106, 65, 161, 215, 231, 246, 6, 106, 65, 12, 119, 32,
-  11, 106, 34, 9, 32, 6, 65, 127, 115, 114, 32, 2, 65, 10, 119, 34, 11, 115,
-  106, 65, 161, 215, 231, 246, 6, 106, 65, 7, 119, 32, 8, 106, 34, 12, 65, 10,
-  119, 34, 3, 106, 32, 3, 32, 25, 32, 9, 65, 10, 119, 34, 2, 106, 32, 2, 32, 27,
-  32, 6, 65, 10, 119, 34, 6, 106, 32, 6, 32, 11, 32, 19, 106, 32, 8, 32, 16,
-  106, 32, 12, 32, 9, 65, 127, 115, 114, 32, 6, 115, 106, 65, 161, 215, 231,
-  246, 6, 106, 65, 5, 119, 32, 11, 106, 34, 6, 32, 2, 113, 32, 12, 32, 2, 65,
-  127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 11, 119, 106,
-  34, 2, 32, 3, 113, 32, 6, 32, 3, 65, 127, 115, 113, 114, 106, 65, 220, 249,
-  238, 248, 120, 106, 65, 12, 119, 106, 34, 9, 32, 6, 65, 10, 119, 34, 3, 113,
-  32, 2, 32, 3, 65, 127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106,
-  65, 14, 119, 106, 34, 11, 32, 2, 65, 10, 119, 34, 2, 113, 32, 9, 32, 2, 65,
-  127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 15, 119, 32, 3,
-  106, 34, 12, 65, 10, 119, 34, 6, 106, 32, 21, 32, 11, 65, 10, 119, 34, 8, 106,
-  32, 16, 32, 9, 65, 10, 119, 34, 9, 106, 32, 2, 32, 10, 106, 32, 3, 32, 23,
-  106, 32, 9, 32, 12, 113, 32, 11, 32, 9, 65, 127, 115, 113, 114, 106, 65, 220,
-  249, 238, 248, 120, 106, 65, 14, 119, 32, 2, 106, 34, 3, 32, 8, 113, 32, 12,
-  32, 8, 65, 127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 15,
-  119, 32, 9, 106, 34, 2, 32, 6, 113, 32, 3, 32, 6, 65, 127, 115, 113, 114, 106,
-  65, 220, 249, 238, 248, 120, 106, 65, 9, 119, 32, 8, 106, 34, 9, 32, 3, 65,
-  10, 119, 34, 3, 113, 32, 2, 32, 3, 65, 127, 115, 113, 114, 106, 65, 220, 249,
-  238, 248, 120, 106, 65, 8, 119, 32, 6, 106, 34, 11, 32, 2, 65, 10, 119, 34, 2,
-  113, 32, 9, 32, 2, 65, 127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120,
-  106, 65, 9, 119, 32, 3, 106, 34, 12, 65, 10, 119, 34, 6, 106, 32, 1, 32, 11,
-  65, 10, 119, 34, 8, 106, 32, 17, 32, 9, 65, 10, 119, 34, 9, 106, 32, 2, 32,
-  20, 106, 32, 3, 32, 28, 106, 32, 9, 32, 12, 113, 32, 11, 32, 9, 65, 127, 115,
-  113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 14, 119, 32, 2, 106, 34,
-  3, 32, 8, 113, 32, 12, 32, 8, 65, 127, 115, 113, 114, 106, 65, 220, 249, 238,
-  248, 120, 106, 65, 5, 119, 32, 9, 106, 34, 2, 32, 6, 113, 32, 3, 32, 6, 65,
-  127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 6, 119, 32, 8,
-  106, 34, 8, 32, 3, 65, 10, 119, 34, 3, 113, 32, 2, 32, 3, 65, 127, 115, 113,
-  114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 8, 119, 32, 6, 106, 34, 9, 32,
-  2, 65, 10, 119, 34, 2, 113, 32, 8, 32, 2, 65, 127, 115, 113, 114, 106, 65,
-  220, 249, 238, 248, 120, 106, 65, 6, 119, 32, 3, 106, 34, 11, 65, 10, 119, 34,
-  12, 106, 32, 23, 32, 9, 65, 10, 119, 34, 6, 106, 32, 21, 32, 8, 65, 10, 119,
-  34, 8, 106, 32, 2, 32, 24, 106, 32, 3, 32, 22, 106, 32, 8, 32, 11, 113, 32, 9,
-  32, 8, 65, 127, 115, 113, 114, 106, 65, 220, 249, 238, 248, 120, 106, 65, 5,
-  119, 32, 2, 106, 34, 3, 32, 6, 113, 32, 11, 32, 6, 65, 127, 115, 113, 114,
-  106, 65, 220, 249, 238, 248, 120, 106, 65, 12, 119, 32, 8, 106, 34, 2, 32, 3,
-  32, 12, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 9,
-  119, 32, 6, 106, 34, 6, 32, 2, 32, 3, 65, 10, 119, 34, 3, 65, 127, 115, 114,
-  115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 15, 119, 32, 12, 106, 34, 8,
-  32, 6, 32, 2, 65, 10, 119, 34, 2, 65, 127, 115, 114, 115, 106, 65, 206, 250,
-  207, 202, 122, 106, 65, 5, 119, 32, 3, 106, 34, 9, 65, 10, 119, 34, 11, 106,
-  32, 24, 32, 8, 65, 10, 119, 34, 12, 106, 32, 16, 32, 6, 65, 10, 119, 34, 6,
-  106, 32, 2, 32, 20, 106, 32, 3, 32, 27, 106, 32, 9, 32, 8, 32, 6, 65, 127,
-  115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 11, 119, 32, 2, 106,
-  34, 3, 32, 9, 32, 12, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207, 202,
-  122, 106, 65, 6, 119, 32, 6, 106, 34, 2, 32, 3, 32, 11, 65, 127, 115, 114,
-  115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 8, 119, 32, 12, 106, 34, 6,
-  32, 2, 32, 3, 65, 10, 119, 34, 3, 65, 127, 115, 114, 115, 106, 65, 206, 250,
-  207, 202, 122, 106, 65, 13, 119, 32, 11, 106, 34, 8, 32, 6, 32, 2, 65, 10,
-  119, 34, 2, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65,
-  12, 119, 32, 3, 106, 34, 9, 65, 10, 119, 34, 11, 106, 32, 10, 32, 8, 65, 10,
-  119, 34, 12, 106, 32, 28, 32, 6, 65, 10, 119, 34, 6, 106, 32, 2, 32, 19, 106,
-  32, 1, 32, 3, 106, 32, 9, 32, 8, 32, 6, 65, 127, 115, 114, 115, 106, 65, 206,
-  250, 207, 202, 122, 106, 65, 5, 119, 32, 2, 106, 34, 3, 32, 9, 32, 12, 65,
-  127, 115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 12, 119, 32, 6,
-  106, 34, 2, 32, 3, 32, 11, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207,
-  202, 122, 106, 65, 13, 119, 32, 12, 106, 34, 6, 32, 2, 32, 3, 65, 10, 119, 34,
-  8, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 14, 119,
-  32, 11, 106, 34, 9, 32, 6, 32, 2, 65, 10, 119, 34, 11, 65, 127, 115, 114, 115,
-  106, 65, 206, 250, 207, 202, 122, 106, 65, 11, 119, 32, 8, 106, 34, 12, 65,
-  10, 119, 34, 33, 32, 4, 106, 32, 1, 32, 10, 32, 27, 32, 23, 32, 21, 32, 23,
-  32, 25, 32, 28, 32, 19, 32, 17, 32, 23, 32, 16, 32, 17, 32, 24, 32, 32, 32,
-  31, 32, 4, 65, 127, 115, 114, 32, 29, 115, 106, 32, 18, 106, 65, 230, 151,
-  138, 133, 5, 106, 65, 8, 119, 32, 30, 106, 34, 3, 65, 10, 119, 34, 2, 106, 32,
-  7, 32, 27, 106, 32, 5, 32, 23, 106, 32, 4, 32, 20, 106, 32, 30, 32, 3, 32, 29,
-  32, 5, 65, 127, 115, 114, 115, 106, 32, 1, 106, 65, 230, 151, 138, 133, 5,
-  106, 65, 9, 119, 32, 4, 106, 34, 4, 32, 3, 32, 7, 65, 127, 115, 114, 115, 106,
-  65, 230, 151, 138, 133, 5, 106, 65, 9, 119, 32, 5, 106, 34, 5, 32, 4, 32, 2,
-  65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 11, 119, 32,
-  7, 106, 34, 7, 32, 5, 32, 4, 65, 10, 119, 34, 4, 65, 127, 115, 114, 115, 106,
-  65, 230, 151, 138, 133, 5, 106, 65, 13, 119, 32, 2, 106, 34, 3, 32, 7, 32, 5,
-  65, 10, 119, 34, 5, 65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5,
-  106, 65, 15, 119, 32, 4, 106, 34, 2, 65, 10, 119, 34, 13, 106, 32, 22, 32, 3,
-  65, 10, 119, 34, 14, 106, 32, 26, 32, 7, 65, 10, 119, 34, 7, 106, 32, 5, 32,
-  21, 106, 32, 4, 32, 25, 106, 32, 2, 32, 3, 32, 7, 65, 127, 115, 114, 115, 106,
-  65, 230, 151, 138, 133, 5, 106, 65, 15, 119, 32, 5, 106, 34, 4, 32, 2, 32, 14,
-  65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 5, 119, 32,
-  7, 106, 34, 5, 32, 4, 32, 13, 65, 127, 115, 114, 115, 106, 65, 230, 151, 138,
-  133, 5, 106, 65, 7, 119, 32, 14, 106, 34, 7, 32, 5, 32, 4, 65, 10, 119, 34, 4,
-  65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 7, 119, 32,
-  13, 106, 34, 3, 32, 7, 32, 5, 65, 10, 119, 34, 5, 65, 127, 115, 114, 115, 106,
-  65, 230, 151, 138, 133, 5, 106, 65, 8, 119, 32, 4, 106, 34, 2, 65, 10, 119,
-  34, 13, 106, 32, 28, 32, 3, 65, 10, 119, 34, 14, 106, 32, 15, 32, 7, 65, 10,
-  119, 34, 7, 106, 32, 5, 32, 19, 106, 32, 4, 32, 10, 106, 32, 2, 32, 3, 32, 7,
-  65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 11, 119, 32,
-  5, 106, 34, 4, 32, 2, 32, 14, 65, 127, 115, 114, 115, 106, 65, 230, 151, 138,
-  133, 5, 106, 65, 14, 119, 32, 7, 106, 34, 5, 32, 4, 32, 13, 65, 127, 115, 114,
-  115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 14, 119, 32, 14, 106, 34, 7, 32,
-  5, 32, 4, 65, 10, 119, 34, 3, 65, 127, 115, 114, 115, 106, 65, 230, 151, 138,
-  133, 5, 106, 65, 12, 119, 32, 13, 106, 34, 2, 32, 7, 32, 5, 65, 10, 119, 34,
-  13, 65, 127, 115, 114, 115, 106, 65, 230, 151, 138, 133, 5, 106, 65, 6, 119,
-  32, 3, 106, 34, 14, 65, 10, 119, 34, 4, 106, 32, 4, 32, 20, 32, 2, 65, 10,
-  119, 34, 5, 106, 32, 5, 32, 28, 32, 7, 65, 10, 119, 34, 7, 106, 32, 7, 32, 13,
-  32, 25, 106, 32, 3, 32, 22, 106, 32, 7, 32, 14, 113, 32, 2, 32, 7, 65, 127,
-  115, 113, 114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 9, 119, 32, 13, 106,
-  34, 7, 32, 5, 113, 32, 14, 32, 5, 65, 127, 115, 113, 114, 106, 65, 164, 162,
-  183, 226, 5, 106, 65, 13, 119, 106, 34, 5, 32, 4, 113, 32, 7, 32, 4, 65, 127,
-  115, 113, 114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 15, 119, 106, 34, 2,
-  32, 7, 65, 10, 119, 34, 4, 113, 32, 5, 32, 4, 65, 127, 115, 113, 114, 106, 65,
-  164, 162, 183, 226, 5, 106, 65, 7, 119, 106, 34, 13, 32, 5, 65, 10, 119, 34,
-  5, 113, 32, 2, 32, 5, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183, 226, 5,
-  106, 65, 12, 119, 32, 4, 106, 34, 14, 65, 10, 119, 34, 7, 106, 32, 1, 32, 13,
-  65, 10, 119, 34, 3, 106, 32, 15, 32, 2, 65, 10, 119, 34, 2, 106, 32, 5, 32,
-  18, 106, 32, 4, 32, 26, 106, 32, 2, 32, 14, 113, 32, 13, 32, 2, 65, 127, 115,
-  113, 114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 8, 119, 32, 5, 106, 34, 4,
-  32, 3, 113, 32, 14, 32, 3, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183,
-  226, 5, 106, 65, 9, 119, 32, 2, 106, 34, 5, 32, 7, 113, 32, 4, 32, 7, 65, 127,
-  115, 113, 114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 11, 119, 32, 3, 106,
-  34, 2, 32, 4, 65, 10, 119, 34, 4, 113, 32, 5, 32, 4, 65, 127, 115, 113, 114,
-  106, 65, 164, 162, 183, 226, 5, 106, 65, 7, 119, 32, 7, 106, 34, 13, 32, 5,
-  65, 10, 119, 34, 5, 113, 32, 2, 32, 5, 65, 127, 115, 113, 114, 106, 65, 164,
-  162, 183, 226, 5, 106, 65, 7, 119, 32, 4, 106, 34, 14, 65, 10, 119, 34, 7,
-  106, 32, 7, 32, 27, 32, 13, 65, 10, 119, 34, 3, 106, 32, 21, 32, 2, 65, 10,
-  119, 34, 2, 106, 32, 5, 32, 16, 106, 32, 4, 32, 10, 106, 32, 2, 32, 14, 113,
-  32, 13, 32, 2, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183, 226, 5, 106,
-  65, 12, 119, 32, 5, 106, 34, 4, 32, 3, 113, 32, 14, 32, 3, 65, 127, 115, 113,
-  114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 7, 119, 32, 2, 106, 34, 5, 32,
-  7, 113, 32, 4, 32, 7, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183, 226, 5,
-  106, 65, 6, 119, 32, 3, 106, 34, 7, 32, 4, 65, 10, 119, 34, 4, 113, 32, 5, 32,
-  4, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 15, 119,
-  106, 34, 3, 32, 5, 65, 10, 119, 34, 5, 113, 32, 7, 32, 5, 65, 127, 115, 113,
-  114, 106, 65, 164, 162, 183, 226, 5, 106, 65, 13, 119, 32, 4, 106, 34, 2, 65,
-  10, 119, 34, 13, 106, 32, 19, 32, 3, 65, 10, 119, 34, 14, 106, 32, 18, 32, 7,
-  65, 10, 119, 34, 7, 106, 32, 5, 32, 17, 106, 32, 4, 32, 24, 106, 32, 2, 32, 7,
-  113, 32, 3, 32, 7, 65, 127, 115, 113, 114, 106, 65, 164, 162, 183, 226, 5,
-  106, 65, 11, 119, 32, 5, 106, 34, 4, 32, 2, 65, 127, 115, 114, 32, 14, 115,
-  106, 65, 243, 253, 192, 235, 6, 106, 65, 9, 119, 32, 7, 106, 34, 5, 32, 4, 65,
-  127, 115, 114, 32, 13, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 7, 119,
-  32, 14, 106, 34, 7, 32, 5, 65, 127, 115, 114, 32, 4, 65, 10, 119, 34, 4, 115,
-  106, 65, 243, 253, 192, 235, 6, 106, 65, 15, 119, 32, 13, 106, 34, 3, 32, 7,
-  65, 127, 115, 114, 32, 5, 65, 10, 119, 34, 5, 115, 106, 65, 243, 253, 192,
-  235, 6, 106, 65, 11, 119, 32, 4, 106, 34, 2, 65, 10, 119, 34, 13, 106, 32, 27,
-  32, 3, 65, 10, 119, 34, 14, 106, 32, 22, 32, 7, 65, 10, 119, 34, 7, 106, 32,
-  1, 32, 5, 106, 32, 4, 32, 20, 106, 32, 2, 32, 3, 65, 127, 115, 114, 32, 7,
-  115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 8, 119, 32, 5, 106, 34, 4, 32,
-  2, 65, 127, 115, 114, 32, 14, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 6,
-  119, 32, 7, 106, 34, 5, 32, 4, 65, 127, 115, 114, 32, 13, 115, 106, 65, 243,
-  253, 192, 235, 6, 106, 65, 6, 119, 32, 14, 106, 34, 7, 32, 5, 65, 127, 115,
-  114, 32, 4, 65, 10, 119, 34, 4, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65,
-  14, 119, 32, 13, 106, 34, 3, 32, 7, 65, 127, 115, 114, 32, 5, 65, 10, 119, 34,
-  5, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 12, 119, 32, 4, 106, 34, 2,
-  65, 10, 119, 34, 13, 106, 32, 15, 32, 3, 65, 10, 119, 34, 14, 106, 32, 24, 32,
-  7, 65, 10, 119, 34, 7, 106, 32, 5, 32, 16, 106, 32, 4, 32, 10, 106, 32, 2, 32,
-  3, 65, 127, 115, 114, 32, 7, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 13,
-  119, 32, 5, 106, 34, 4, 32, 2, 65, 127, 115, 114, 32, 14, 115, 106, 65, 243,
-  253, 192, 235, 6, 106, 65, 5, 119, 32, 7, 106, 34, 5, 32, 4, 65, 127, 115,
-  114, 32, 13, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 14, 119, 32, 14,
-  106, 34, 7, 32, 5, 65, 127, 115, 114, 32, 4, 65, 10, 119, 34, 4, 115, 106, 65,
-  243, 253, 192, 235, 6, 106, 65, 13, 119, 32, 13, 106, 34, 3, 32, 7, 65, 127,
-  115, 114, 32, 5, 65, 10, 119, 34, 5, 115, 106, 65, 243, 253, 192, 235, 6, 106,
-  65, 13, 119, 32, 4, 106, 34, 2, 65, 10, 119, 34, 13, 106, 32, 22, 32, 3, 65,
-  10, 119, 34, 14, 106, 32, 10, 32, 7, 65, 10, 119, 34, 7, 106, 32, 5, 32, 26,
-  106, 32, 4, 32, 21, 106, 32, 2, 32, 3, 65, 127, 115, 114, 32, 7, 115, 106, 65,
-  243, 253, 192, 235, 6, 106, 65, 7, 119, 32, 5, 106, 34, 5, 32, 2, 65, 127,
-  115, 114, 32, 14, 115, 106, 65, 243, 253, 192, 235, 6, 106, 65, 5, 119, 32, 7,
-  106, 34, 10, 32, 5, 113, 32, 13, 32, 10, 65, 127, 115, 113, 114, 106, 65, 233,
-  237, 181, 211, 7, 106, 65, 15, 119, 32, 14, 106, 34, 4, 32, 10, 113, 32, 5,
-  65, 10, 119, 34, 7, 32, 4, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181,
-  211, 7, 106, 65, 5, 119, 32, 13, 106, 34, 5, 32, 4, 113, 32, 10, 65, 10, 119,
-  34, 3, 32, 5, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65,
-  8, 119, 32, 7, 106, 34, 10, 65, 10, 119, 34, 2, 106, 32, 17, 32, 5, 65, 10,
-  119, 34, 13, 106, 32, 25, 32, 4, 65, 10, 119, 34, 4, 106, 32, 3, 32, 28, 106,
-  32, 10, 32, 7, 32, 19, 106, 32, 5, 32, 10, 113, 32, 4, 32, 10, 65, 127, 115,
-  113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 11, 119, 32, 3, 106, 34,
-  10, 113, 32, 13, 32, 10, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211,
-  7, 106, 65, 14, 119, 32, 4, 106, 34, 4, 32, 10, 113, 32, 2, 32, 4, 65, 127,
-  115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 14, 119, 32, 13, 106,
-  34, 5, 32, 4, 113, 32, 10, 65, 10, 119, 34, 7, 32, 5, 65, 127, 115, 113, 114,
-  106, 65, 233, 237, 181, 211, 7, 106, 65, 6, 119, 32, 2, 106, 34, 10, 32, 5,
-  113, 32, 4, 65, 10, 119, 34, 3, 32, 10, 65, 127, 115, 113, 114, 106, 65, 233,
-  237, 181, 211, 7, 106, 65, 14, 119, 32, 7, 106, 34, 4, 65, 10, 119, 34, 2,
-  106, 32, 26, 32, 10, 65, 10, 119, 34, 13, 106, 32, 24, 32, 5, 65, 10, 119, 34,
-  5, 106, 32, 3, 32, 16, 106, 32, 7, 32, 18, 106, 32, 4, 32, 10, 113, 32, 5, 32,
-  4, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 6, 119,
-  32, 3, 106, 34, 10, 32, 4, 113, 32, 13, 32, 10, 65, 127, 115, 113, 114, 106,
-  65, 233, 237, 181, 211, 7, 106, 65, 9, 119, 32, 5, 106, 34, 4, 32, 10, 113,
-  32, 2, 32, 4, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65,
-  12, 119, 32, 13, 106, 34, 5, 32, 4, 113, 32, 10, 65, 10, 119, 34, 7, 32, 5,
-  65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 9, 119, 32,
-  2, 106, 34, 10, 32, 5, 113, 32, 4, 65, 10, 119, 34, 3, 32, 10, 65, 127, 115,
-  113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 12, 119, 32, 7, 106, 34, 4,
-  65, 10, 119, 34, 2, 32, 17, 106, 32, 1, 32, 5, 65, 10, 119, 34, 5, 106, 32, 3,
-  32, 15, 106, 32, 7, 32, 20, 106, 32, 4, 32, 10, 113, 32, 5, 32, 4, 65, 127,
-  115, 113, 114, 106, 65, 233, 237, 181, 211, 7, 106, 65, 5, 119, 32, 3, 106,
-  34, 1, 32, 4, 113, 32, 10, 65, 10, 119, 34, 4, 32, 1, 65, 127, 115, 113, 114,
-  106, 65, 233, 237, 181, 211, 7, 106, 65, 15, 119, 32, 5, 106, 34, 10, 32, 1,
-  113, 32, 2, 32, 10, 65, 127, 115, 113, 114, 106, 65, 233, 237, 181, 211, 7,
-  106, 65, 8, 119, 32, 4, 106, 34, 5, 32, 10, 65, 10, 119, 34, 7, 115, 32, 4,
-  32, 16, 106, 32, 10, 32, 1, 65, 10, 119, 34, 1, 115, 32, 5, 115, 106, 65, 8,
-  119, 32, 2, 106, 34, 10, 115, 106, 65, 5, 119, 32, 1, 106, 34, 16, 65, 10,
-  119, 34, 4, 106, 32, 19, 32, 5, 65, 10, 119, 34, 19, 106, 32, 1, 32, 15, 106,
-  32, 10, 32, 19, 115, 32, 16, 115, 106, 65, 12, 119, 32, 7, 106, 34, 1, 32, 4,
-  115, 32, 7, 32, 21, 106, 32, 16, 32, 10, 65, 10, 119, 34, 15, 115, 32, 1, 115,
-  106, 65, 9, 119, 32, 19, 106, 34, 10, 115, 106, 65, 12, 119, 32, 15, 106, 34,
-  19, 32, 10, 65, 10, 119, 34, 21, 115, 32, 15, 32, 18, 106, 32, 10, 32, 1, 65,
-  10, 119, 34, 1, 115, 32, 19, 115, 106, 65, 5, 119, 32, 4, 106, 34, 15, 115,
-  106, 65, 14, 119, 32, 1, 106, 34, 18, 65, 10, 119, 34, 10, 106, 32, 19, 65,
-  10, 119, 34, 19, 32, 24, 106, 32, 1, 32, 20, 106, 32, 15, 32, 19, 115, 32, 18,
-  115, 106, 65, 6, 119, 32, 21, 106, 34, 1, 32, 10, 115, 32, 21, 32, 22, 106,
-  32, 18, 32, 15, 65, 10, 119, 34, 15, 115, 32, 1, 115, 106, 65, 8, 119, 32, 19,
-  106, 34, 18, 115, 106, 65, 13, 119, 32, 15, 106, 34, 20, 32, 18, 65, 10, 119,
-  34, 19, 115, 32, 15, 32, 26, 106, 32, 18, 32, 1, 65, 10, 119, 34, 1, 115, 32,
-  20, 115, 106, 65, 6, 119, 32, 10, 106, 34, 15, 115, 106, 65, 5, 119, 32, 1,
-  106, 34, 18, 65, 10, 119, 34, 10, 106, 54, 2, 8, 32, 0, 32, 1, 32, 23, 106,
-  32, 15, 32, 20, 65, 10, 119, 34, 1, 115, 32, 18, 115, 106, 65, 15, 119, 32,
-  19, 106, 34, 20, 65, 10, 119, 34, 24, 32, 31, 32, 8, 32, 22, 106, 32, 12, 32,
-  9, 32, 6, 65, 10, 119, 34, 22, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207,
-  202, 122, 106, 65, 8, 119, 32, 11, 106, 34, 23, 65, 10, 119, 106, 106, 54, 2,
-  4, 32, 0, 32, 29, 32, 11, 32, 17, 106, 32, 23, 32, 12, 32, 9, 65, 10, 119, 34,
-  21, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207, 202, 122, 106, 65, 5, 119,
-  32, 22, 106, 34, 17, 106, 32, 19, 32, 28, 106, 32, 18, 32, 15, 65, 10, 119,
-  34, 15, 115, 32, 20, 115, 106, 65, 13, 119, 32, 1, 106, 34, 18, 65, 10, 119,
-  106, 54, 2, 0, 32, 0, 32, 1, 32, 27, 106, 32, 10, 32, 20, 115, 32, 18, 115,
-  106, 65, 11, 119, 32, 15, 106, 34, 1, 32, 21, 32, 32, 106, 32, 22, 32, 26,
-  106, 32, 17, 32, 23, 32, 33, 65, 127, 115, 114, 115, 106, 65, 206, 250, 207,
-  202, 122, 106, 65, 6, 119, 106, 106, 54, 2, 16, 32, 0, 32, 21, 32, 30, 106,
-  32, 10, 106, 32, 15, 32, 25, 106, 32, 18, 32, 24, 115, 32, 1, 115, 106, 65,
-  11, 119, 106, 54, 2, 12, 11, 36, 1, 2, 127, 65, 160, 8, 33, 1, 2, 64, 2, 64,
-  2, 64, 32, 0, 14, 2, 1, 0, 2, 11, 65, 192, 8, 33, 1, 11, 32, 1, 33, 2, 11, 32,
-  2, 11, 6, 0, 32, 0, 16, 1, 11, 11, 27, 1, 0, 65, 128, 8, 11, 20, 1, 35, 69,
-  103, 137, 171, 205, 239, 254, 220, 186, 152, 118, 84, 50, 16, 240, 225, 210,
-  195, 0, 38, 9, 112, 114, 111, 100, 117, 99, 101, 114, 115, 1, 12, 112, 114,
-  111, 99, 101, 115, 115, 101, 100, 45, 98, 121, 1, 5, 99, 108, 97, 110, 103, 6,
-  49, 49, 46, 48, 46, 48,
-];
+import WASM from "./wasm.mjs";
 
 /**
  * Performs the message digest algorithm RIPEMD160 on input data.
